refactor(tests): share LLM-core fake MXQ backends across logits_to_keep tests
Extract the LLM-core fakes (StaticLastOnlyMxq, DynamicAxisMxq, FakeBackend,
FakeInputBufferInfo, make_model) into tests/transformers/_fake_mxq.py so
test_logits_to_keep.py imports them instead of redefining. The dual-input
Qwen3-VL variants stay local — different infer signature (no batch_params)
and two input buffers — with a comment explaining why.

Pure test refactor; no behavior changes.

pytest tail:
  tests\transformers\test_logits_to_keep.py .............................. [ 71%]
  .                                                                        [ 73%]
  tests\transformers\image_text_to_text\test_qwen3_vl_logits_to_keep.py .. [ 78%]
  .........                                                                [100%]
  ======================= 42 passed, 2 warnings in 0.32s ========================
