+++ tests/transformers/test_logits_to_keep.py
@@ -22,128 +22,11 @@ import pytest
 import torch
 
 from mblt_model_zoo.hf_transformers.utils.modeling_utils import MobilintModelMixin
-
-
-# ---------------------------------------------------------------------------
-# Fake MXQ + backend helpers
-# ---------------------------------------------------------------------------
-
-
-class _FakeInputBufferInfo:
-    def __init__(self, max_width: int, max_cache_size: int = 128):
-        self.max_width = max_width
-        self.max_cache_size = max_cache_size
-
-
-class _StaticLastOnlyMxq:
-    """MXQ stub that emits only the last-token logits per infer call.
-
-    ``infer`` accepts both the single-item shape (``[chunk], None, cache_size``)
-    and the batched shape (``[chunk], None, 0, batch_params``). Every call is
-    recorded so tests can assert chunk boundaries.
-    """
-
-    def __init__(self, vocab_size: int = 5, max_width: int = 4):
-        self.vocab_size = vocab_size
-        self.max_width = max_width
-        self.calls: list[dict] = []
-        self._batched_counter = 0
-
-    def get_input_buffer_info(self):
-        return [_FakeInputBufferInfo(max_width=self.max_width)]
-
-    def get_model_output_shape(self):
-        # Static token axis (second-to-last); "1" signals last-token-only output.
-        return [(1, 1, self.vocab_size)]
-
-    def infer(self, inputs, _extra, cache_size, batch_params=None):
-        chunk = np.asarray(inputs[0])
-        if batch_params is None:
-            self.calls.append({"shape": tuple(chunk.shape), "cache_size": int(cache_size)})
-            return [np.full((1, 1, self.vocab_size), fill_value=float(cache_size), dtype=np.float32)]
-
-        self._batched_counter += 1
-        active_batch = len(batch_params)
-        payload = np.stack(
-            [
-                np.full((self.vocab_size,), fill_value=float(p.cache_id * 100 + self._batched_counter), dtype=np.float32)
-                for p in batch_params
-            ],
-            axis=0,
-        )
-        self.calls.append(
-            {
-                "shape": tuple(chunk.shape),
-                "batch": [(p.cache_id, p.sequence_length, p.cache_size) for p in batch_params],
-            }
-        )
-        assert payload.shape == (active_batch, self.vocab_size)
-        return [payload]
-
-
-class _DynamicAxisMxq:
-    """MXQ stub that emits per-position logits."""
-
-    def __init__(self, vocab_size: int = 5, max_width: int = 4):
-        self.vocab_size = vocab_size
-        self.max_width = max_width
-        self.calls: list[dict] = []
-        self._batched_counter = 0
-
-    def get_input_buffer_info(self):
-        return [_FakeInputBufferInfo(max_width=self.max_width)]
-
-    def get_model_output_shape(self):
-        # Dynamic token axis; the shared helper interprets this as all-logits.
-        return [(1, -1, self.vocab_size)]
-
-    def infer(self, inputs, _extra, cache_size, batch_params=None):
-        chunk = np.asarray(inputs[0])
-        if batch_params is None:
-            # Single-item: chunk is (1, 1, chunk_len, hidden). Emit
-            # (1, chunk_len, vocab) so the token axis is second-to-last.
-            chunk_len = int(chunk.shape[2])
-            offset = int(cache_size) * self.vocab_size
-            values = np.arange(chunk_len * self.vocab_size, dtype=np.float32) + offset
-            self.calls.append(
-                {"shape": tuple(chunk.shape), "cache_size": int(cache_size), "chunk_len": chunk_len}
-            )
-            return [values.reshape(1, chunk_len, self.vocab_size)]
-
-        self._batched_counter += 1
-        total_tokens = sum(int(p.sequence_length) for p in batch_params)
-        flat = np.arange(total_tokens * self.vocab_size, dtype=np.float32).reshape(
-            total_tokens, self.vocab_size
-        )
-        # Encode per-item id into the payload so batched tests can spot mixups.
-        offset = 0
-        for p in batch_params:
-            flat[offset : offset + int(p.sequence_length), :] += p.cache_id * 1000.0
-            offset += int(p.sequence_length)
-        self.calls.append(
-            {
-                "shape": tuple(chunk.shape),
-                "batch": [(p.cache_id, p.sequence_length, p.cache_size) for p in batch_params],
-            }
-        )
-        return [flat]
-
-
-class _FakeBackend:
-    def __init__(self, mxq_model):
-        self.mxq_model = mxq_model
-
-
-def _make_model(mxq, *, max_batch_size: int = 1) -> MobilintModelMixin:
-    """Construct a bare ``MobilintModelMixin`` without triggering NPU init."""
-    model = MobilintModelMixin.__new__(MobilintModelMixin)
-    model.npu_backend = _FakeBackend(mxq)
-    config_kwargs = {"npu_prefill_chunk_size": None}
-    if max_batch_size > 1:
-        config_kwargs["max_batch_size"] = max_batch_size
-    model.config = type("Config", (), config_kwargs)()
-    model.npu_time = None
-    return model
+from tests.transformers._fake_mxq import (
+    DynamicAxisMxq,
+    StaticLastOnlyMxq,
+    make_model,
+)
 
 
 # ---------------------------------------------------------------------------
@@ -194,16 +77,16 @@ class TestNormalizeLogitsToKeep:
 
 class TestMxqSupportsAllLogits:
     def test_static_token_axis_returns_false(self) -> None:
-        model = _make_model(_StaticLastOnlyMxq())
+        model = make_model(StaticLastOnlyMxq())
         assert model._mxq_supports_all_logits() is False
 
     def test_dynamic_token_axis_returns_true(self) -> None:
-        model = _make_model(_DynamicAxisMxq())
+        model = make_model(DynamicAxisMxq())
         assert model._mxq_supports_all_logits() is True
 
     def test_result_is_cached_on_instance(self) -> None:
-        mxq = _StaticLastOnlyMxq()
-        model = _make_model(mxq)
+        mxq = StaticLastOnlyMxq()
+        model = make_model(mxq)
         assert model._mxq_supports_all_logits() is False
 
         # Flip the underlying shape; a fresh probe would now claim dynamic.
@@ -212,11 +95,11 @@ class TestMxqSupportsAllLogits:
         assert model._mxq_supports_all_logits() is False
 
     def test_raising_backend_falls_back_to_false(self) -> None:
-        class _ExplodingMxq(_StaticLastOnlyMxq):
+        class _ExplodingMxq(StaticLastOnlyMxq):
             def get_model_output_shape(self):  # noqa: D401
                 raise RuntimeError("no shape for you")
 
-        model = _make_model(_ExplodingMxq())
+        model = make_model(_ExplodingMxq())
         assert model._mxq_supports_all_logits() is False
 
 
@@ -235,7 +118,7 @@ class TestLlmForwardSingle:
         logits_to_keep,
         prefill_chunk_size: Optional[int] = None,
     ):
-        model = _make_model(mxq)
+        model = make_model(mxq)
         inputs_embeds = torch.arange(seq_len * hidden_size, dtype=torch.float32).reshape(
             1, seq_len, hidden_size
         )
@@ -250,7 +133,7 @@ class TestLlmForwardSingle:
         return model, logits
 
     def test_default_keep_uses_fast_path_with_last_token_logits(self) -> None:
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
         _model, logits = self._run(mxq, seq_len=6, hidden_size=3, logits_to_keep=1, prefill_chunk_size=3)
 
         # Two chunks of size 3 with the caller's prefill_chunk_size.
@@ -260,7 +143,7 @@ class TestLlmForwardSingle:
         assert logits.shape == (1, mxq.vocab_size)
 
     def test_default_keep_ignores_dynamic_axis_probe(self) -> None:
-        mxq = _DynamicAxisMxq(vocab_size=5, max_width=4)
+        mxq = DynamicAxisMxq(vocab_size=5, max_width=4)
         model, logits = self._run(mxq, seq_len=6, hidden_size=3, logits_to_keep=1, prefill_chunk_size=3)
 
         # is_default_keep short-circuits before probing, so no cached probe result yet.
@@ -270,7 +153,7 @@ class TestLlmForwardSingle:
         assert logits.shape == (3, mxq.vocab_size)
 
     def test_dynamic_axis_keep_all_returns_every_position(self) -> None:
-        mxq = _DynamicAxisMxq(vocab_size=5, max_width=4)
+        mxq = DynamicAxisMxq(vocab_size=5, max_width=4)
         model, logits = self._run(
             mxq, seq_len=6, hidden_size=3, logits_to_keep=0, prefill_chunk_size=3
         )
@@ -295,14 +178,14 @@ class TestLlmForwardSingle:
         np.testing.assert_allclose(logits.numpy(), expected)
 
     def test_dynamic_axis_keep_last_n_slices_expected_positions(self) -> None:
-        mxq = _DynamicAxisMxq(vocab_size=5, max_width=4)
+        mxq = DynamicAxisMxq(vocab_size=5, max_width=4)
         _model, logits = self._run(
             mxq, seq_len=6, hidden_size=3, logits_to_keep=2, prefill_chunk_size=3
         )
         assert logits.shape == (2, mxq.vocab_size)
 
     def test_dynamic_axis_tensor_indices_pick_out_positions(self) -> None:
-        mxq = _DynamicAxisMxq(vocab_size=5, max_width=4)
+        mxq = DynamicAxisMxq(vocab_size=5, max_width=4)
         indices = torch.tensor([0, 2, 5])
         _model, logits = self._run(
             mxq, seq_len=6, hidden_size=3, logits_to_keep=indices, prefill_chunk_size=3
@@ -310,7 +193,7 @@ class TestLlmForwardSingle:
         assert logits.shape == (3, mxq.vocab_size)
 
     def test_fallback_interleaves_size_one_infer_for_kept_positions(self) -> None:
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
         indices = torch.tensor([2, 5])
         _model, logits = self._run(
             mxq, seq_len=6, hidden_size=3, logits_to_keep=indices, prefill_chunk_size=4
@@ -326,8 +209,8 @@ class TestLlmForwardSingle:
     def test_fallback_advances_kv_cache_across_all_positions(self) -> None:
         from mblt_model_zoo.hf_transformers.utils.cache_utils import MobilintCache
 
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
-        model = _make_model(mxq)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        model = make_model(mxq)
         cache = MobilintCache(mxq)
         seq_len = 6
         inputs_embeds = torch.zeros(1, seq_len, 3)
@@ -346,7 +229,7 @@ class TestLlmForwardSingle:
 
     def test_fallback_empty_tensor_returns_empty_logits(self) -> None:
         """An empty ``logits_to_keep`` tensor must not raise on np.concatenate."""
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
         _model, logits = self._run(
             mxq,
             seq_len=6,
@@ -362,7 +245,7 @@ class TestLlmForwardSingle:
 
     def test_fallback_all_out_of_range_indices_returns_empty_logits(self) -> None:
         """All-out-of-range ``logits_to_keep`` must not raise on np.concatenate."""
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
         _model, logits = self._run(
             mxq,
             seq_len=6,
@@ -377,8 +260,8 @@ class TestLlmForwardSingle:
     def test_fallback_empty_tensor_still_advances_kv_cache(self) -> None:
         from mblt_model_zoo.hf_transformers.utils.cache_utils import MobilintCache
 
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
-        model = _make_model(mxq)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        model = make_model(mxq)
         cache = MobilintCache(mxq)
         seq_len = 6
         inputs_embeds = torch.zeros(1, seq_len, 3)
@@ -413,7 +296,7 @@ class TestLlmForwardBatch:
         prefill_chunk_size: Optional[int] = None,
         max_batch_size: int = 4,
     ):
-        model = _make_model(mxq, max_batch_size=max_batch_size)
+        model = make_model(mxq, max_batch_size=max_batch_size)
         batch, seq_len = attention_mask.shape
         inputs_embeds = torch.arange(batch * seq_len * hidden_size, dtype=torch.float32).reshape(
             batch, seq_len, hidden_size
@@ -430,7 +313,7 @@ class TestLlmForwardBatch:
         return model, logits
 
     def test_default_keep_path_preserves_existing_shape(self) -> None:
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=2)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=2)
         attention_mask = torch.tensor([[1, 1, 0], [1, 1, 1]], dtype=torch.long)
         _model, logits = self._run(
             mxq, attention_mask=attention_mask, hidden_size=4, logits_to_keep=1, prefill_chunk_size=2
@@ -438,7 +321,7 @@ class TestLlmForwardBatch:
         assert logits.shape == (2, 1, mxq.vocab_size)
 
     def test_dynamic_axis_keep_all_returns_per_item_padded_stack(self) -> None:
-        mxq = _DynamicAxisMxq(vocab_size=5, max_width=4)
+        mxq = DynamicAxisMxq(vocab_size=5, max_width=4)
         attention_mask = torch.tensor([[1, 1, 0], [1, 1, 1]], dtype=torch.long)
         _model, logits = self._run(
             mxq,
@@ -454,7 +337,7 @@ class TestLlmForwardBatch:
         assert torch.all(torch.isinf(logits[0, 2]) & (logits[0, 2] < 0))
 
     def test_dynamic_axis_tensor_indices_apply_per_item(self) -> None:
-        mxq = _DynamicAxisMxq(vocab_size=5, max_width=4)
+        mxq = DynamicAxisMxq(vocab_size=5, max_width=4)
         attention_mask = torch.tensor([[1, 1, 1, 1], [1, 1, 1, 0]], dtype=torch.long)
         indices = torch.tensor([0, 2])
         _model, logits = self._run(
@@ -468,7 +351,7 @@ class TestLlmForwardBatch:
         assert logits.shape == (2, 2, mxq.vocab_size)
 
     def test_fallback_interleaves_prefill_and_size_one_across_batch(self) -> None:
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
         attention_mask = torch.tensor([[1, 1, 1, 1], [1, 1, 1, 1]], dtype=torch.long)
         # Request positions 2 and 3 so the fallback branch runs; logits_to_keep=1
         # would hit the fast path and skip the fallback entirely.
@@ -487,7 +370,7 @@ class TestLlmForwardBatch:
         assert logits.shape == (2, 2, mxq.vocab_size)
 
     def test_fallback_handles_variable_per_item_kept_positions(self) -> None:
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
         # Item 0 has 3 real tokens; item 1 has 4. Keep-all on both yields
         # 3 and 4 positions respectively → padding on item 0.
         attention_mask = torch.tensor([[1, 1, 1, 0], [1, 1, 1, 1]], dtype=torch.long)
@@ -511,7 +394,7 @@ class TestLlmForwardBatch:
         until position 5. The per-item pointer lets the middle walk use the
         caller's ``prefill_chunk_size``.
         """
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
         # Item 0 seq_len=6 keeps {0, 5}; item 1 seq_len=1 keeps {0} (5 clamped away).
         attention_mask = torch.tensor(
             [[1, 1, 1, 1, 1, 1], [1, 0, 0, 0, 0, 0]], dtype=torch.long
@@ -542,7 +425,7 @@ class TestLlmForwardBatch:
         after cursor=1, even though item 0 was inactive from cursor=2 onward
         and item 1 didn't need another capture until 5.
         """
-        mxq = _StaticLastOnlyMxq(vocab_size=5, max_width=4)
+        mxq = StaticLastOnlyMxq(vocab_size=5, max_width=4)
         # Item 0 seq_len=2 keeps {1}; item 1 seq_len=6 keeps {5} via last-index.
         attention_mask = torch.tensor(
             [[1, 1, 0, 0, 0, 0], [1, 1, 1, 1, 1, 1]], dtype=torch.long
